requirements.txt
aiofiles==24.1.0
altair==6.0.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
gradio==6.11.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.19.0
numpy==2.4.4
ollama==0.6.1
orjson==3.11.8
packaging==26.0
pandas==3.0.2
pefile==2024.8.26
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.20.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
pytz==2026.1.post1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
